+++ network/src/archive/challenge.rs
@@ -7,7 +7,7 @@ use tape_client::{
     get_block_account, get_miner_account, get_epoch_account, get_tape_account
 };
 
-use crate::store::TapeStore;
+use crate::store::*;
 use super::queue::Tx;
 use super::sync::sync_segments_from_solana;
