+++ network/src/store/mod.rs
@@ -1,13 +1,24 @@
 pub mod consts;
 pub mod error;
-mod sector;
 mod layout;
+mod sector;
+mod segment;
 mod tape_store;
+mod health;
+mod tape;
+mod merkle;
+mod stats;
 mod helpers;
 
 pub use consts::*;
 pub use error::StoreError;
 pub use tape_store::TapeStore;
+pub use health::{StoreStaticKeys, HealthOps};
+pub use tape::TapeOps;
+pub use segment::SegmentOps;
+pub use sector::{Sector, SectorOps};
+pub use merkle::MerkleOps;
+pub use stats::{LocalStats, StatsOps};
 pub use helpers::{
     primary,
     secondary_mine,
