+++ network/src/store/sector.rs
@@ -1,6 +1,7 @@
-use bytemuck::{Pod, Zeroable};
+use solana_sdk::pubkey::Pubkey;
+use bytemuck::{Pod, Zeroable, try_from_bytes, bytes_of};
 use tape_api::consts::*;
-use super::consts::*;
+use super::{consts::*, TapeStore, error::StoreError, layout::ColumnFamily};
 
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug)]
@@ -43,5 +44,40 @@ impl Sector {
     pub fn count_segments(&self) -> usize {
         let bitmap_len = SECTOR_LEAVES / 8;
         self.0[..bitmap_len].iter().map(|byte| byte.count_ones() as usize).sum()
+    }
+}
+
+pub trait SectorOps {
+    fn get_sector(&self, tape_address: &Pubkey, sector_number: u64) -> Result<Sector, StoreError>;
+    fn put_sector(&self, tape_address: &Pubkey, sector_number: u64, sector: &Sector) -> Result<(), StoreError>;
+}
+
+impl SectorOps for TapeStore {
+    fn get_sector(&self, tape_address: &Pubkey, sector_number: u64) -> Result<Sector, StoreError> {
+        let cf = self.get_cf_handle(ColumnFamily::Sectors)?;
+        let mut key = Vec::with_capacity(TAPE_STORE_SLOTS_KEY_SIZE);
+        key.extend_from_slice(&tape_address.to_bytes());
+        key.extend_from_slice(&sector_number.to_be_bytes());
+        
+        let data = self
+            .db
+            .get_cf(&cf, &key)?
+            .ok_or_else(|| StoreError::SegmentNotFoundForAddress(tape_address.to_string(), sector_number))?;
+        
+        if data.len() != SECTOR_HEADER_BYTES + SECTOR_LEAVES * PACKED_SEGMENT_SIZE {
+            return Err(StoreError::InvalidSectorSize(data.len()));
+        }
+        
+        Ok(*try_from_bytes(&data).map_err(|_| StoreError::InvalidSectorSize(data.len()))?)
+    }
+
+    fn put_sector(&self, tape_address: &Pubkey, sector_number: u64, sector: &Sector) -> Result<(), StoreError> {
+        let cf = self.get_cf_handle(ColumnFamily::Sectors)?;
+        let mut key = Vec::with_capacity(TAPE_STORE_SLOTS_KEY_SIZE);
+        key.extend_from_slice(&tape_address.to_bytes());
+        key.extend_from_slice(&sector_number.to_be_bytes());
+        
+        self.db.put_cf(&cf, &key, bytes_of(sector))?;
+        Ok(())
     }
 }
