+++ network/src/store/tape.rs
@@ -1,0 +1,51 @@
+use solana_sdk::pubkey::Pubkey;
+use rocksdb::WriteBatch;
+use super::{ TapeStore, error::StoreError, layout::ColumnFamily};
+use crate::metrics::inc_total_tapes_written;
+
+pub trait TapeOps {
+    fn put_tape_address(&self, tape_number: u64, address: &Pubkey) -> Result<(), StoreError>;
+    fn get_tape_number(&self, address: &Pubkey) -> Result<u64, StoreError>;
+    fn get_tape_address(&self, tape_number: u64) -> Result<Pubkey, StoreError>;
+}
+
+impl TapeOps for TapeStore {
+    fn put_tape_address(&self, tape_number: u64, address: &Pubkey) -> Result<(), StoreError> {
+        let cf_tape_by_number = self.get_cf_handle(ColumnFamily::TapeByNumber)?;
+        let cf_tape_by_address = self.get_cf_handle(ColumnFamily::TapeByAddress)?;
+        let tape_number_key = tape_number.to_be_bytes().to_vec();
+        let address_key = address.to_bytes().to_vec();
+        let mut batch = WriteBatch::default();
+        batch.put_cf(&cf_tape_by_number, &tape_number_key, address.to_bytes());
+        batch.put_cf(&cf_tape_by_address, &address_key, tape_number.to_be_bytes());
+        self.db.write(batch)?;
+        inc_total_tapes_written();
+        Ok(())
+    }
+
+    fn get_tape_number(&self, address: &Pubkey) -> Result<u64, StoreError> {
+        let cf = self.get_cf_handle(ColumnFamily::TapeByAddress)?;
+        let key = address.to_bytes().to_vec();
+        let tape_number_bytes = self
+            .db
+            .get_cf(&cf, &key)?
+            .ok_or_else(|| StoreError::ValueNotFoundForAddress(address.to_string()))?;
+        Ok(u64::from_be_bytes(
+            tape_number_bytes
+                .try_into()
+                .map_err(|_| StoreError::InvalidSegmentKey)?,
+        ))
+    }
+
+    fn get_tape_address(&self, tape_number: u64) -> Result<Pubkey, StoreError> {
+        let cf = self.get_cf_handle(ColumnFamily::TapeByNumber)?;
+        let key = tape_number.to_be_bytes().to_vec();
+        let address_bytes = self
+            .db
+            .get_cf(&cf, &key)?
+            .ok_or(StoreError::TapeNotFound(tape_number))?;
+
+        Pubkey::try_from(address_bytes.as_slice())
+            .map_err(|e| StoreError::InvalidPubkey(e.to_string()))
+    }
+}
