+++ network/src/utils/health.rs
@@ -1,5 +1,5 @@
 use log::debug;
-use crate::store::TapeStore;
+use crate::store::*;
 
 pub fn drift_status(store: &TapeStore, latest_slot: u64, last_processed_slot: u64) {
     // TODO: this function is not working right.
